Remove launchd based installer
This apprach was nice in some ways, but is flaky and doesn't work as
reliabily when using an installer package.

Addtionally, it isn't possible to install a per-version build service
with the environment variable approach.

This approach moves the XCBBuildService inside of Xcode to a new
location and symlinks the BKBuildService to the default location.

It dynamically uses the `.default` build service in the case it hasn't
been installed just yet ( e.g. testing )