[Stability] Fix BEPStream.FileHandle not closing on occasion
According to
https://developer.apple.com/documentation/foundation/filehandle/1413393-closefile
FileHandle will both close it's self and tear it down. I've observed the
file not closing and `readabilityHandler` going into an infinite loop,
even when `deinit` is called. This happens only sometimes, but when it
does, it will really consume resources.

Mitigations include disabling it at the end of the stream and in deinit.